Django-Middleware-0x03/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3